backend/requirements_perfect.txt
# Perfect Portfolio API - Production Dependencies
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Validation & Serialization
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Optional - External APIs
alpaca-py==0.12.1
twelvedata==1.2.14
yfinance==0.2.28
openai==1.3.7

# Optional - Production
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
